+++ usaspending_api/awards/v2/filters/sub_award.py
@@ -2,12 +2,12 @@ import itertools
 import logging
 
 from django.db.models import Q
-from usaspending_api.accounts.helpers import build_tas_codes_filter
 from usaspending_api.awards.models_matviews import SubawardView
 from usaspending_api.awards.v2.filters.filter_helpers import combine_date_range_queryset, total_obligation_queryset
 from usaspending_api.awards.v2.filters.location_filter_geocode import geocode_filter_locations
 from usaspending_api.common.exceptions import InvalidParameterException
 from usaspending_api.references.models import PSC
+from usaspending_api.search.helpers import build_tas_codes_filter
 from usaspending_api.search.v2 import elasticsearch_helper
 from usaspending_api.settings import API_MAX_DATE, API_MIN_DATE, API_SEARCH_MIN_DATE
